Starting to get more serious now. Went in and cleaned up a bunch of code structures to improve clarity and communicate intention (I didn't have to use the Boolean function on all the ticket age context (hours, mins, secs) but it makes the code really obvious). Further cleanup was done on variable declarations and I've tidied up the grouping of constructs.
Proper ticketlist presence checking (To make sure uberTones doesn't
run on any page that's not a ticketlist) was also implemented.

Started the beginnings (configuration object and asset objects) of the
data structures that the playTones function will operate with, as a
lot will be going on before before that function is called and within
that function, it's important to get the program flow properly sorted
out before beginning the implementation.